GithubContributionモデルを作った
